+++ triangle_test.cpp
@@ -10,11 +10,11 @@ void make_octahedron(std::list <geometry_object> *l_t,
                      const vec &v1, const vec &v2, const vec &v3,
                      const point &p) {
     assert(l_t != nullptr);
-    int n;
+    unsigned long long n;
     point p1(p.x + v1.x, p.y + v1.y, p.z + v1.z);
     point p2(p.x + v2.x, p.y + v2.y, p.z + v2.z);
     point p3(p.x + v3.x, p.y + v3.y, p.z + v3.z);
-    n = (int) l_t->size();
+    n = l_t->size();
     triangle t(p1, p2, p3);
     geometry_object o(TRIANGLE, t, n);
     l_t->push_back(o);
@@ -22,7 +22,7 @@ void make_octahedron(std::list <geometry_object> *l_t,
     p1 = point(p.x - v1.x, p.y + v1.y, p.z + v1.z);
     p2 = point(p.x - v2.x, p.y + v2.y, p.z + v2.z);
     p3 = point(p.x - v3.x, p.y + v3.y, p.z + v3.z);
-    n = (int) l_t->size();
+    n = l_t->size();
     t = triangle(p1, p2, p3);
     o = geometry_object(TRIANGLE, t, n);
     l_t->push_back(o);
@@ -30,7 +30,7 @@ void make_octahedron(std::list <geometry_object> *l_t,
     p1 = point(p.x + v1.x, p.y - v1.y, p.z + v1.z);
     p2 = point(p.x + v2.x, p.y - v2.y, p.z + v2.z);
     p3 = point(p.x + v3.x, p.y - v3.y, p.z + v3.z);
-    n = (int) l_t->size();
+    n = l_t->size();
     t = triangle(p1, p2, p3);
     o = geometry_object(TRIANGLE, t, n);
     l_t->push_back(o);
@@ -38,7 +38,7 @@ void make_octahedron(std::list <geometry_object> *l_t,
     p1 = point(p.x + v1.x, p.y + v1.y, p.z - v1.z);
     p2 = point(p.x + v2.x, p.y + v2.y, p.z - v2.z);
     p3 = point(p.x + v3.x, p.y + v3.y, p.z - v3.z);
-    n = (int) l_t->size();
+    n = l_t->size();
     t = triangle(p1, p2, p3);
     o = geometry_object(TRIANGLE, t, n);
     l_t->push_back(o);
@@ -46,7 +46,7 @@ void make_octahedron(std::list <geometry_object> *l_t,
     p1 = point(p.x - v1.x, p.y - v1.y, p.z + v1.z);
     p2 = point(p.x - v2.x, p.y - v2.y, p.z + v2.z);
     p3 = point(p.x - v3.x, p.y - v3.y, p.z + v3.z);
-    n = (int) l_t->size();
+    n = l_t->size();
     t = triangle(p1, p2, p3);
     o = geometry_object(TRIANGLE, t, n);
     l_t->push_back(o);
@@ -54,7 +54,7 @@ void make_octahedron(std::list <geometry_object> *l_t,
     p1 = point(p.x - v1.x, p.y + v1.y, p.z - v1.z);
     p2 = point(p.x - v2.x, p.y + v2.y, p.z - v2.z);
     p3 = point(p.x - v3.x, p.y + v3.y, p.z - v3.z);
-    n = (int) l_t->size();
+    n = l_t->size();
     t = triangle(p1, p2, p3);
     o = geometry_object(TRIANGLE, t, n);
     l_t->push_back(o);
@@ -62,7 +62,7 @@ void make_octahedron(std::list <geometry_object> *l_t,
     p1 = point(p.x + v1.x, p.y - v1.y, p.z - v1.z);
     p2 = point(p.x + v2.x, p.y - v2.y, p.z - v2.z);
     p3 = point(p.x + v3.x, p.y - v3.y, p.z - v3.z);
-    n = (int) l_t->size();
+    n = l_t->size();
     t = triangle(p1, p2, p3);
     o = geometry_object(TRIANGLE, t, n);
     l_t->push_back(o);
@@ -70,7 +70,7 @@ void make_octahedron(std::list <geometry_object> *l_t,
     p1 = point(p.x - v1.x, p.y - v1.y, p.z - v1.z);
     p2 = point(p.x - v2.x, p.y - v2.y, p.z - v2.z);
     p3 = point(p.x - v3.x, p.y - v3.y, p.z - v3.z);
-    n = (int) l_t->size();
+    n = l_t->size();
     t = triangle(p1, p2, p3);
     o = geometry_object(TRIANGLE, t, n);
     l_t->push_back(o);
@@ -82,7 +82,7 @@ void make_triangle_stack(std::list <geometry::geometry_object> *l_t,
     assert(l_t != nullptr);
     assert(step > 0);
     triangle t(t_base.p1_ret(), t_base.p2_ret(), t_base.p3_ret());
-    geometry_object o(TRIANGLE, t, (int) l_t->size());
+    geometry_object o(TRIANGLE, t, l_t->size());
     l_t->push_back(o);
     plane pl = t.make_plane();
     vec n(pl.a, pl.b, pl.c);
@@ -117,7 +117,7 @@ void make_triangle_stack(std::list <geometry::geometry_object> *l_t,
         p3.y += n.y;
         p3.z += n.z;
         t = triangle(p1, p2, p3);
-        o = geometry_object(TRIANGLE, t, (int) l_t->size());
+        o = geometry_object(TRIANGLE, t, l_t->size());
         l_t->push_back(o);
     }
 }
@@ -153,7 +153,7 @@ void test2(std::list <geometry_object> *l_t) {
     p2 = point(1, 2, 0);
     p3 = point(1, 1, 0.03);
     t = triangle(p1, p2, p3);
-    geometry_object o(TRIANGLE, t, (int) l_t->size());
+    geometry_object o(TRIANGLE, t, l_t->size());
     l_t->push_back(o);
 }
 #endif
@@ -169,13 +169,13 @@ void test3(std::list <geometry_object> *l_t) {
     p2 = point(1, 2, 0);
     p3 = point(1, 1, 0.03);
     t = triangle(p1, p2, p3);
-    geometry_object o(TRIANGLE, t, (int) l_t->size());
+    geometry_object o(TRIANGLE, t, l_t->size());
     l_t->push_back(o);
     p1 = point(1, 1, 0.5);
     p2 = point(1, 2, 0.45);
     p3 = point(1, 1, 0.66);
     t = triangle(p1, p2, p3);
-    o = geometry_object(geometry::TRIANGLE, t, (int) l_t->size());
+    o = geometry_object(geometry::TRIANGLE, t, l_t->size());
     l_t->push_back(o);
 }
 #endif
@@ -218,12 +218,12 @@ void test6(std::list <geometry_object> *l_t) {
     triangle t(p1, p2, p3);
     make_triangle_stack(l_t, t, 0.01, 1);
     p1 = point (0, 0, 0);
-    geometry_object o(geometry::POINT, p1, (int) l_t->size());
+    geometry_object o(geometry::POINT, p1, l_t->size());
     l_t->push_back(o);
     p1 = point(-1, -1, -1);
     p2 = point(-2, -2, -2);
     cut c(p1, p2);
-    o = geometry_object(geometry::CUT, c, (int) l_t->size());
+    o = geometry_object(geometry::CUT, c, l_t->size());
     l_t->push_back(o);
 }
 #endif
@@ -234,15 +234,15 @@ void test7(std::list <geometry_object> *l_t) {
     point p2(10, 0, 0);
     point p3(0, 10, 0);
     triangle t(p1, p2, p3);
-    geometry_object o(geometry::TRIANGLE, t, (int) l_t->size());
+    geometry_object o(geometry::TRIANGLE, t, l_t->size());
     l_t->push_back(o);
     p1 = point (1, 1, 1);
-    o = geometry_object(geometry::POINT, p1, (int) l_t->size());
+    o = geometry_object(geometry::POINT, p1, l_t->size());
     l_t->push_back(o);
     p1 = point(0, 0, 0);
     p2 = point(1, 1, 0);
     cut c(p1, p2);
-    o = geometry_object(geometry::CUT, c, (int) l_t->size());
+    o = geometry_object(geometry::CUT, c, l_t->size());
     l_t->push_back(o);
 }
 #endif
@@ -253,15 +253,15 @@ void test8(std::list <geometry_object> *l_t) {
     point p2(10, 0, 0);
     point p3(0, 10, 0);
     triangle t(p1, p2, p3);
-    geometry_object o(geometry::TRIANGLE, t, (int) l_t->size());
+    geometry_object o(geometry::TRIANGLE, t, l_t->size());
     l_t->push_back(o);
     p1 = point (1, 1, 1);
-    o = geometry_object(geometry::POINT, p1, (int) l_t->size());
+    o = geometry_object(geometry::POINT, p1, l_t->size());
     l_t->push_back(o);
     p1 = point(2, 2, 1);
     p2 = point(0, 0, 1);
     cut c(p1, p2);
-    o = geometry_object(geometry::CUT, c, (int) l_t->size());
+    o = geometry_object(geometry::CUT, c, l_t->size());
     l_t->push_back(o);
 }
 #endif
